FrontEnd/requirements.txt
gitpython
ipython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
psutil
PyYAML>=5.3.1
requests>=2.23.0

datasets==2.*
fastapi==0.74.*
sentencepiece==0.1.*
transformers==4.*
uvicorn[standard]==0.17.*

ffmpeg-python
python-multipart

streamlit
Pillow
redis